Feature/category

+++ src/main/java/com/depromeet/team5/domain/Store.java
@@ -15,9 +15,9 @@ public class Store {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Float latitude;
+    private Double latitude;
 
-    private Float longitude;
+    private Double longitude;
 
     private String storeName;
