+++ src/main/java/com/depromeet/team5/service/CategoryService.java
@@ -4,6 +4,6 @@ import com.depromeet.team5.domain.CategoryTypes;
 import com.depromeet.team5.dto.CategoryDto;
 
 public interface CategoryService {
-    CategoryDto getDistanceList(Float latitude, Float longitude, CategoryTypes category);
-    CategoryDto getReviewList(Float latitude, Float longitude,CategoryTypes category);
+    CategoryDto getDistanceList(Double latitude, Double longitude, CategoryTypes category);
+    CategoryDto getReviewList(Double latitude, Double longitude,CategoryTypes category);
 }
